fix: handle `bytes=0-` correctly in objectnode
Some clients (e.g. docker registry) send a get to fetch the whole file
by specifying `bytes=0-` as the range. In this case, ChubaoFS will set
the range to 0-0 and the `Content-Length` to 1 (upper - lower + 1). It
will then error out when writing the object, since it will write more
than the specified `Content-Length`.

As a simple fix, we can treat a rangeUpper of 0 the same as max, which
should result in correct behavior.

Signed-off-by: Belhorma Bendebiche <amro@bndb.sh>