fdstore: Add fdstore
fdstore is a tool to help live-upgrade cfs-client. It has several options:

 -p <port>: cfs-client pprof port
 -s: suspend cfs-client
 -r: resume cfs-client
 -d <nodes file>,<handles file>: dump files
 -n: generate dynamic UDS file name
 -v: show version

If suspend is specified, it will show the unix domain socket file. The
UDS file is used to do cfs-client resotre. If -n is not used, the UDS
name is always "/tmp/ChubaoFS-fdstore.sock". Example of suspend looks
like:

	fdstore -p 17410 -s

Signed-off-by: shengyong002 <shengyong002@ke.com>