datanode: Add deletePartition to delete orphan datapartition
When creating a new datapartition:

	        OpCreateDataPartition
	master -----------------------> datanode1
	        |---------XXX---------> datanode2
	        `---------------------> datanode3

Both datanode1 and datanode2 create dp and response to master
successfully. But timeout occurs when datanode2 is creating dp.
Master will send an OpDeleteDataPartition to the three datanodes.
Both datanode1 and datanode2 delete the dp successfully.

Unfortunately, datanode2 does not receive the delete request since
timeout of creating connection or datanode2 finds that the dp does
not exist because previous create request does not finish yet.

After datanode2 finishes creating the dp, the dp only has itself
alone. Let's call this kind of dp as orphan dp.

Orphan dp will try to establish a correct raft status with datanode1
and datanode3. Obviously, it fails every time. However, there is no
way to clear the orphan dp.

So this patch adds a new API for datanode. It will try to delete
such orphan dp by:

	curl '127.0.0.1:17320/deletePartition?id=[PartitionID]&vol=[VolumeName]'

NOTE that it must be VERY CAREFULL to call this API, datanode only
checks whether master has metadata of this dp. If not, datanode
will delete the dp directly.

Signed-off-by: shengyong002 <shengyong002@ke.com>