libsdk: Introduce UserTask
UserTask is a sub-structure in client. It is only used when libsdk
works as a daemon for now. In preload case, the topology looks like:

	            C U B E F S    S E R V E R S
	                     ^         ^
	                     |         |
	 Preload Daemon [ client1  client2  ... ]
	                     /\         |
	                    /  \        |
	                task1 task2   task
         (libsdk)          |    |       |
	------------------------------------------
         (user)       userapp userapp userapp

where the client in Preload Daemon acts as a proxy, which is in
charge of communicating with cubefs servers. Multiple tasks could
be attached to one client. Userapp sends commands and requests to
its own task.