fix: disable fsyncOnClose still flush the data
When fsyncOnClose is disabled, it implies that the end user knows the
written data will not be persistent upon closing the file descriptor.
The only guarentee to persist data is to invoke fsync system call
explicitly.

Without this commit, closing a file descriptor will still flush the data
which is not as expected.

Signed-off-by: Shuoran Liu <shuoranliu@gmail.com>