CI: Fix large chunk of whitespace in ltp result
There are lots of large chunk of whitespace in ltp result. That
makes the result output diffault to read. The whitespace is caused
by truncating /tmp/ltp.log and redirection it to a temporary file.

This patch simplies the execution of runltp and simplies the checking
of ltp log file.

Signed-off-by: Sheng Yong <shengyong2021@gmail.com>