metanode: Introduce rocksdb associated with column families
This patch introduces rocksdb to metapartition. Each metapartition
could has a rocksdb, which is used to persist Inode, Dentry, Extend
and Multipart. The life cycle of a rocksdb is the same as its mp.
In this way, the original snapshot directory could be replaced by
rocksdb.

Each rocksdb has 5 column families:
 * default: is not used but required by rocksdb
 * inode: persists all Inodes
 * dentry: persists all Dentries
 * extend: persists all Extends
 * multipart: persists all Multiparts

Signed-off-by: Sheng Yong <shengyong2021@gmail.com>